+++ CHANGELOG.md
@@ -6,11 +6,25 @@
 
 ### OKIT-Desktop Version 0.3.0
 #### Features
+1. Unpdate to indicate this is the final release.
 #### Bug Fixes
 
 ### OKIT-Classic Version 0.70.0
 ##### Features
+1. Rebrand to Edge Cloud Design Tool
+2. Add feature to allow the user to specify that the Identity Resource should be split from non-identity resources and create multiple directories.
+3. Option to Edit Lock resource, i.e. do not allow editting of fields if the lock is closed.
+4. Save now always guarentees a ".okit" extension
+5. Add Simple Theme framework.
+6. Create Reducer based Contexts in readiness to improve code.
+    1. Cache
+    2. Console Config
+    3. Theme
+    4. Active File
+7. Restructure build workflow to create Classic & Desktop releases.
 ### Bug Fixes
+1. Fix security list options issue where the fields were not being correctly hidden.
+2. Fix issue where Documentation tab was not always highlighted correctly.
 
 
 ## Version 0.69.1